Action Bar (WIP)
Refactor item instance to use DatabaseLink
Labels now pop up when aiming at loot
Yeeted HUD trigger group UI into the abyss
Created rebindable action bar (missing: assigning actions)
Created Consumable Items
Refactored behaviors to work on consumable items
Implemented Velocity Conversion Behavior

+++ Assets/Shaders/Raymarching/VolumeRender.shader
@@ -155,7 +155,8 @@
 	
 	float4 RayMarch( in float3 origin, in float3 direction, in float zbuf, in float2 screenUV, out float scatterSum )
 	{
-        half rand = frac(tex2D(_DitheringTex, screenUV * _DitheringCoords.xy).r + _FrameNumber * 1.61803398875);
+        float rand = frac(tex2D(_DitheringTex, screenUV * _DitheringCoords.xy).r + _FrameNumber * 1.61803398875);
+        //half rand = frac(nrand(screenUV) + _FrameNumber * 1.61803398875);
 		//half rand = nrand(screenUV + frac(_Time.x)) * 2;
 		float4 sum = (float4)0.;
 		scatterSum = 0.;
